third_party/blink/web_tests/wpt_internal/cobalt/HTMLVideoElementExtension/getAndSetMaxVideoCapabilities.html

<!DOCTYPE html>
<html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script type='module'>
    test(function(){
    var video = document.createElement('video');
    assert_equals('', video.maxVideoCapabilities, 'maxVideoCapabilities can retrieve unset value');
    }, 'maxVideoCapabilities for HTMLVideoElement');

    test(function(){
    var video = document.createElement('video');
    video.src = 'testStringToSetSRC';
    // a try catch must be used here, as trying to use assert_throws_dom or assert_throws_js
    // will not work due to the returned error immediately stopping the test.
    try{
      video.maxVideoCapabilities = 'width=1920; height=1080; framerate=15;';
      throw AssertionError('setMaxVideoCapabilities successfully ran when not supposed to');
    }
    catch ({name, message}){
      assert_equals("InvalidStateError", name, 'InvalidStateError is thrown.');
      assert_equals("Failed to set the 'maxVideoCapabilities' property on 'HTMLVideoElement': Cannot set maximum capabilities after src is defined.", message,
                    'Error message matches defined error message in setMaxVideoCapabilities');
    }
    }, 'setMaxVideoCapabilities cannot be set after src is defined');

    test(function(){
    var video = document.createElement('video');
    video.maxVideoCapabilities = 'width=1920; height=1080; framerate=15;';
    assert_equals('width=1920; height=1080; framerate=15;', video.maxVideoCapabilities,
                  'setMaxVideoCapabilities properly updates maxVideoCapabilities');
    }, 'setMaxVideoCapabilities successfully updates HTMLVideoElement');
  </script>
</body>
</html>
